requirements.txt
discord.py==1.7.3
matplotlib==3.4.2
numpy==1.21.1
pandas==1.3.3
ccxt==1.58.9
